video_analysis_project/backend/requirements.txt
fastapi
uvicorn
ffmpeg-python
whisper
ultralytics
opencv-python-headless
numpy
pandas
